requirements.txt
pdfplumber == 0.11.5
python-dateutil == 2.9.0.post0
google-api-python-client == 2.157.0
google-auth-httplib2 == 0.2.0
google-auth-oauthlib == 1.2.1